refactor: split unilab.tools into owner layers and scripts/tools
- scripts/tools/import_robot.py: one-off URDF->MJCF asset authoring
  (hardcodes REPO_ROOT, drops unilab-import-robot console entry)
- scripts/motion/bones_seed_csv.py: CSV contract helper co-located with
  its only consumers
- src/unilab/utils/nan_viz.py: viewer next to its nan_guard dump producer
- src/unilab/assets/pull.py: thin CLI over assets.hub
- src/unilab/base/backend/mujoco/export_scene.py: MuJoCo-only export
- src/unilab/visualization/teaser.py: motrix teaser scene renderer
- src/unilab/cli_completion.py: shell completion sibling of cli.py
- update console entries (pyproject.toml, pyproject.rocm.toml), demo.py,
  shell completion fallbacks, tests, docs; drop api_reference/tools page
