--- docs/sphinx/source/api_reference/tools/index.md
@@ -1,15 +1,0 @@
-# `unilab.tools` — CLI Tools
-
-Console-script entrypoints registered in `pyproject.toml`:
-
-- `unilab-viz-nan` — interactive NaN trace viewer for failed runs.
-- `unilab-export-scene` — dump the resolved scene of a task to disk.
-
-```{eval-rst}
-.. autosummary::
-   :toctree: _autosummary
-   :template: autosummary/module.rst
-   :recursive:
-
-   unilab.tools
-```
