+++ docs/sphinx/source/en/2-user_guide/7-tooling/4-scene_export.md
@@ -1,6 +1,6 @@
 # Scene Export
 
-Scene export is implemented by `src/unilab/tools/export_scene.py` and registered
+Scene export is implemented by `src/unilab/base/backend/mujoco/export_scene.py` and registered
 as the `unilab-export-scene` console entry in `pyproject.toml`. It accepts a
 MuJoCo XML or MJB model path, writes `scene.xml`, copies mesh assets when they
 are discoverable, and can create a zip archive.
