+++ docs/sphinx/source/zh_CN/2-user_guide/7-tooling/3-nan_visualizer.md
@@ -8,4 +8,4 @@ uv run train --algo ppo --task go2_joystick_flat --sim mujoco \
   training.nan_guard.output_dir=/tmp/unilab/nan_dumps
 ```
 
-viewer 的实现是 `src/unilab/tools/viz_nan.py`，注册为 `unilab-viz-nan` 控制台入口。它会回放一个 dump 路径，并让你选择环境索引。dump 格式和往返加载由 `tests/test_nan_guard.py` 覆盖。
+viewer 的实现是 `src/unilab/utils/nan_viz.py`，注册为 `unilab-viz-nan` 控制台入口。它会回放一个 dump 路径，并让你选择环境索引。dump 格式和往返加载由 `tests/test_nan_guard.py` 覆盖。
