+++ docs/sphinx/source/zh_CN/2-user_guide/7-tooling/5-robot_import.md
@@ -25,7 +25,7 @@ task/reward/env 语义。
 如果只有 URDF，使用仓库自带脚本进行转换：
 
 ```bash
-uv run unilab-import-robot <urdf_path> [robot_name]
+uv run scripts/tools/import_robot.py <urdf_path> [robot_name]
 ```
 
 ```{important}
@@ -58,7 +58,7 @@ visual mesh 作为 collision mesh，尽量把碰撞体简化为 box / capsule / 
 
 ## 输出产物
 
-使用 `uv run unilab-import-robot <urdf_path> [robot_name]` 转换后，会在仓库内生成：
+使用 `uv run scripts/tools/import_robot.py <urdf_path> [robot_name]` 转换后，会在仓库内生成：
 
 - `src/unilab/assets/robots/<robot_name>/assets/`：转换并整理后的 mesh 资产。
 - `src/unilab/assets/robots/<robot_name>/<robot_name>.xml`：机器人 MJCF 描述，只包含机器人
