+++ scripts/completions/unilab.bash
@@ -14,7 +14,7 @@ _unilab_uv_complete() {
     local candidates
     if ! mapfile -t candidates < <(
         uv run --no-sync unilab-complete --cword "$COMP_CWORD" -- "${COMP_WORDS[@]}" 2>/dev/null \
-            || PYTHONPATH="$repo_root/src${PYTHONPATH:+:$PYTHONPATH}" uv run --no-sync python -m unilab.tools.completion --cword "$COMP_CWORD" -- "${COMP_WORDS[@]}" 2>/dev/null
+            || PYTHONPATH="$repo_root/src${PYTHONPATH:+:$PYTHONPATH}" uv run --no-sync python -m unilab.cli_completion --cword "$COMP_CWORD" -- "${COMP_WORDS[@]}" 2>/dev/null
     ); then
         return 0
     fi
