+++ scripts/motion/bones_seed_csv.py
