+++ src/unilab/demo.py
@@ -238,13 +238,13 @@ def _run_teaser_demo() -> int:
     if platform.system() == "Darwin" and Path(sys.executable).name != "mxpython":
         command = [
             _mxpython_executable(),
-            str(_repo_root() / "src" / "unilab" / "tools" / "render_teaser.py"),
+            str(_repo_root() / "src" / "unilab" / "visualization" / "teaser.py"),
         ]
         env = os.environ.copy()
         env["UV_PROJECT_ENVIRONMENT"] = str(_repo_root() / ".venv")
         return subprocess.run(command, check=False, env=env).returncode
 
-    from unilab.tools.render_teaser import main as render_teaser_main
+    from unilab.visualization.teaser import main as render_teaser_main
 
     render_teaser_main()
     return 0
