+++ tests/utils/test_utils_package_policy.py
@@ -9,6 +9,7 @@ ALLOWED_UTILS_MODULES = {
     "device",
     "geometry",
     "nan_guard",
+    "nan_viz",
     "rotation",
     "support_matrix",
     "tensor",
